hotfix: 修正server中路徑無資料夾時的錯誤
調整項目：
1. 在對路徑操作前，先藉由for或exists去確認有此路徑

issue #257

+++ home/views.py
@@ -25,14 +25,15 @@ class InitializeView(View):
     def get(self, request, *arg, **kwargs):
         username = request.user.get_username()
         finish = False
-        for path in ['upload', 'output']:
-            for method in ['DPView', 'k_Anonymity', 'l_Diversity', 't_Closeness']:
-                temp_path = path+'/'+method+'/'+username+'/'
+        for path in [settings.UPLOAD_ROOT, settings.OUTPUT_ROOT]:
+            for method in os.listdir(path):
+                temp_path = path+method+'/'+username+'/'
                 for directory_path in os.listdir(temp_path):
                     shutil.rmtree(temp_path+directory_path)
         path = settings.DPVIEW_TEMP_ROOT+username+'/'
-        for directory_path in os.listdir(path):
-            shutil.rmtree(path+directory_path)
+        if os.path.exists(path):
+            for directory_path in os.listdir(path):
+                shutil.rmtree(path+directory_path)
         if ExecuteModel.objects.filter(user_name=username).exists():
             file = ExecuteModel.objects.get(user_name=username)
             file.delete()
